requirements.txt
fastapi==0.95.2
sqlalchemy==2.0.20
alembic==1.11.1
python-dotenv==1.0.0
requests==2.31.0
faker==19.5.0
pydantic==1.10.8
